feat: rewrite sd-dialog documentation
